+++ src/worker/core/phase/newPhaseDraft.ts
@@ -34,6 +34,8 @@ const newPhaseDraft = async (conditions: Conditions): Promise<PhaseReturn> => {
 
 	await Promise.all(promises);
 
+	await draft.genPlayers(g.get("season"));
+
 	if (g.get("draftType") !== "freeAgents") {
 		// Run lottery only if it hasn't been done yet
 		const draftLotteryResult = await idb.getCopy.draftLotteryResults({